requirements.txt
openai==1.23.1
python-dotenv==1.0.1
